+++ packages/ai/src/types/requests.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { TypedSchema } from '../requests/schema-builder';
+import { ObjectSchema, TypedSchema } from '../requests/schema-builder';
 import { Content, Part } from './content';
 import {
   LanguageModelCreateOptions,
@@ -28,7 +28,7 @@ import {
   HarmCategory,
   ResponseModality
 } from './enums';
-import { ObjectSchemaInterface, SchemaRequest } from './schema';
+import { ObjectSchemaRequest, SchemaRequest } from './schema';
 
 /**
  * Base parameters for a number of methods.
@@ -103,7 +103,7 @@ export interface GenerationConfig {
    * value can be a class generated with a {@link Schema} static method
    * like `Schema.string()` or `Schema.object()` or it can be a plain
    * JS object matching the {@link SchemaRequest} interface.
-   * <br/>Note: This only applies when the specified `responseMIMEType` supports a schema; currently
+   * <br/>Note: This only applies when the specified `responseMimeType` supports a schema; currently
    * this is limited to `application/json` and `text/x.enum`.
    */
   responseSchema?: TypedSchema | SchemaRequest;
@@ -117,6 +117,10 @@ export interface GenerationConfig {
    * @beta
    */
   responseModalities?: ResponseModality[];
+  /**
+   * Configuration for "thinking" behavior of compatible Gemini models.
+   */
+  thinkingConfig?: ThinkingConfig;
 }
 
 /**
@@ -169,7 +173,7 @@ export interface RequestOptions {
  * Defines a tool that model can call to access external knowledge.
  * @public
  */
-export declare type Tool = FunctionDeclarationsTool;
+export type Tool = FunctionDeclarationsTool | GoogleSearchTool;
 
 /**
  * Structured representation of a function declaration as defined by the
@@ -180,7 +184,7 @@ export declare type Tool = FunctionDeclarationsTool;
  * as a Tool by the model and executed by the client.
  * @public
  */
-export declare interface FunctionDeclaration {
+export interface FunctionDeclaration {
   /**
    * The name of the function to call. Must start with a letter or an
    * underscore. Must be a-z, A-Z, 0-9, or contain underscores and dashes, with
@@ -197,8 +201,43 @@ export declare interface FunctionDeclaration {
    * format. Reflects the Open API 3.03 Parameter Object. Parameter names are
    * case-sensitive. For a function with no parameters, this can be left unset.
    */
-  parameters?: ObjectSchemaInterface;
+  parameters?: ObjectSchema | ObjectSchemaRequest;
 }
+
+/**
+ * A tool that allows a Gemini model to connect to Google Search to access and incorporate
+ * up-to-date information from the web into its responses.
+ *
+ * Important: If using Grounding with Google Search, you are required to comply with the
+ * "Grounding with Google Search" usage requirements for your chosen API provider: {@link https://ai.google.dev/gemini-api/terms#grounding-with-google-search | Gemini Developer API}
+ * or Vertex AI Gemini API (see {@link https://cloud.google.com/terms/service-terms | Service Terms}
+ * section within the Service Specific Terms).
+ *
+ * @public
+ */
+export interface GoogleSearchTool {
+  /**
+   * Specifies the Google Search configuration.
+   * Currently, this is an empty object, but it's reserved for future configuration options.
+   * Specifies the Google Search configuration. Currently, this is an empty object, but it's
+   * reserved for future configuration options.
+   *
+   * When using this feature, you are required to comply with the "Grounding with Google Search"
+   * usage requirements for your chosen API provider: {@link https://ai.google.dev/gemini-api/terms#grounding-with-google-search | Gemini Developer API}
+   * or Vertex AI Gemini API (see {@link https://cloud.google.com/terms/service-terms | Service Terms}
+   * section within the Service Specific Terms).
+   */
+  googleSearch: GoogleSearch;
+}
+
+/**
+ * Specifies the Google Search configuration.
+ *
+ * @remarks Currently, this is an empty object, but it's reserved for future configuration options.
+ *
+ * @public
+ */
+export interface GoogleSearch {}
 
 /**
  * A `FunctionDeclarationsTool` is a piece of code that enables the system to
@@ -206,7 +245,7 @@ export declare interface FunctionDeclaration {
  * outside of knowledge and scope of the model.
  * @public
  */
-export declare interface FunctionDeclarationsTool {
+export interface FunctionDeclarationsTool {
   /**
    * Optional. One or more function declarations
    * to be passed to the model along with the current user query. Model may
@@ -269,3 +308,28 @@ export type InferenceMode =
   | 'prefer_on_device'
   | 'only_on_device'
   | 'only_in_cloud';
+
+/*
+ * Configuration for "thinking" behavior of compatible Gemini models.
+ *
+ * Certain models utilize a thinking process before generating a response. This allows them to
+ * reason through complex problems and plan a more coherent and accurate answer.
+ *
+ * @public
+ */
+export interface ThinkingConfig {
+  /**
+   * The thinking budget, in tokens.
+   *
+   * This parameter sets an upper limit on the number of tokens the model can use for its internal
+   * "thinking" process. A higher budget may result in higher quality responses for complex tasks
+   * but can also increase latency and cost.
+   *
+   * If you don't specify a budget, the model will determine the appropriate amount
+   * of thinking based on the complexity of the prompt.
+   *
+   * An error will be thrown if you set a thinking budget for a model that does not support this
+   * feature or if the specified budget is not within the model's supported range.
+   */
+  thinkingBudget?: number;
+}